requirements.txt
logging
streambar
skorch
xgboost
numpy
pandas
pytorch
sklearn
seaborn
matplotlib
scipy
autopep8
docformatter
black
flake8